refactor: 下载交易账单API和下载资金账单现在是公共接口
