+++ docs/changelog.md
@@ -1,3 +1,9 @@
+## 1.0.2.RELEASE
+- 微信支付
+  - feat: 接入微信支付分
+  - feat: 接入微信支付先享卡
+  - fix: 支付回调参数不全的问题
+
 ## 1.0.1.RELEASE
 
 - 微信支付
