+++ docs/quick_start.md
@@ -4,7 +4,7 @@
 <dependency>
     <groupId>cn.felord</groupId>
     <artifactId>payment-spring-boot-starter</artifactId>
-    <version>1.0.1.RELEASE</version>
+    <version>1.0.3.RELEASE</version>
 </dependency>
 ```
 > 基于 **Spring Boot 2.4.1**
