+++ payment-spring-boot-autoconfigure/src/main/java/cn/felord/payment/wechat/enumeration/WechatPayV3Type.java
@@ -44,6 +44,18 @@ public enum WechatPayV3Type {
      * @since 1.0.0.RELEASE
      */
     FILE_DOWNLOAD(HttpMethod.GET, "%s/v3/billdownload/file"),
+    /**
+     * 申请交易账单API.
+     *
+     * @since 1.0.3.RELEASE
+     */
+    TRADEBILL(HttpMethod.GET, "%s/v3/bill/tradebill"),
+    /**
+     * 申请资金账单API.
+     *
+     * @since 1.0.3.RELEASE
+     */
+    FUNDFLOWBILL(HttpMethod.GET, "%s/v3/bill/fundflowbill"),
 
     //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -133,18 +145,6 @@ public enum WechatPayV3Type {
      * @since 1.0.0.RELEASE
      */
     COMBINE_CLOSE(HttpMethod.POST, "%s/v3/combine-transactions/out-trade-no/{combine_out_trade_no}/close"),
-    /**
-     * 申请交易账单API.
-     *
-     * @since 1.0.3.RELEASE
-     */
-    COMBINE_TRADEBILL(HttpMethod.POST, "%s/v3/bill/tradebill"),
-    /**
-     * 申请资金账单API.
-     *
-     * @since 1.0.3.RELEASE
-     */
-    COMBINE_FUNDFLOWBILL(HttpMethod.POST, "%s/v3/bill/fundflowbill"),
 
     //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
